refactor(modulePreloadPolyfill): move polyfill code to own file
